Intention d'achat pour une publicité de bonbons

Ces données simulées reproduisent une étude réalisée au Tech3Lab, dans le cadre de laquelle des cobayes devaient naviguer sur un site internet qui contenait, entre autres choses, une publicité pour des bonbons. Pendant la navigation, un oculomètre mesurait l'endroit où se posait le regard du sujet. On a ainsi pu mesurer si le sujet a regardé la publicité et combien de temps il l’a regardé. De plus, un logiciel d’analyse des expressions faciales (FaceReader) a également été utilisé pour mesurer l’émotion du sujet pendant qu’il regardait la publicité. À la fin de l’expérience, un questionnaire mesurait l’intention d’achat du sujet pour ces bonbons, ainsi que des variables socio-démographiques. Seuls les 120 sujets qui ont regardé la publicité sont inclus dans les données.

intention

Format

une base de donnée de 120 observations contenant les variables suivantes

intention

variable discrète entre 2 et 14; plus elle est élevée, plus le sujet exprime l’intention d’acheter ce produit. Le score a été construit en additionnant les réponses de deux questions sur une échelle de Likert allant de fortement en désaccord (1) à fortement en accord (7)

fixation

durée totale de fixation de la publicité (en secondes).

emotion

une mesure de la valence durant la fixation, soit le ratio de la probabilité d’une émotion positive sur la probabilité d’une émotion négative

sexe

sexe du sujet, soit homme (0) ou femme (1).

age

âge (en années).

revenu

variable catégorique indiquant le revenu annuel du sujet; un parmi (1) [0, 20 000]; (2) [20 000, 60 000] ou (3) 60 000 et plus.

educ

variable catégorique indiquant le niveau d’éducation le plus élevé obtenu, soit (1) secondaire ou moindre; (2) collégial, ou (3) universitaire.

statut

statut matrimonial, soit célibaire (0) ou en couple (1).

achat

variable binaire indiquant si la personne a acheté des bonbons suite au visionnement, oui (1) ou non (0).

nachat

nombre de paquets de bonbons achetés

Source

Marc Fredette